chess-link/lambda/requirements.txt
boto3==1.34.69
python-chess==1.999
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0